wip: remove nft module, rename outgoing transfer dependencies
