+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -10,7 +10,6 @@ import 'package:cryptoplease/features/add_funds/presentation/routes.dart';
 import 'package:cryptoplease/features/app_lock/presentation/routes.dart';
 import 'package:cryptoplease/features/backup_phrase/presentation/routes.dart';
 import 'package:cryptoplease/features/legal/legal_document_view.dart';
-import 'package:cryptoplease/features/nft/presentation/nft_details_screen.dart';
 import 'package:cryptoplease/features/outgoing_transfer/presentation/outgoing_transfer_flow/outgoing_transfer_flow.dart';
 import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/routes.dart';
 import 'package:cryptoplease/features/profile/edit_profile_screen.dart';
@@ -41,7 +40,6 @@ const authenticatedFlowRoutes = AutoRoute<void>(
       ],
     ),
     ...backupPhraseRoutes,
-    AutoRoute<void>(page: NftDetailsScreen),
     AutoRoute<void>(page: TransactionsScreen),
     AutoRoute<void>(page: TransactionDetailsScreen),
     AutoRoute<void>(page: SwapTokenFlowScreen, children: swapTokenRoutes),
