--- packages/cryptoplease/lib/features/nft/bl/nft_collection/event.dart
@@ -1,6 +1,0 @@
-part of 'bloc.dart';
-
-@freezed
-class NftCollectionEvent with _$NftCollectionEvent {
-  const factory NftCollectionEvent.updated() = NftCollectionEventUpdated;
-}
