--- packages/cryptoplease/lib/features/nft/bl/nft_collection/state.dart
@@ -1,10 +1,0 @@
-part of 'bloc.dart';
-
-@freezed
-class NftCollectionState with _$NftCollectionState {
-  const factory NftCollectionState({
-    @Default(ProcessingStateNone()) ProcessingState processingState,
-    @Default(IListConst<NonFungibleToken>([]))
-        IList<NonFungibleToken> nftCollection,
-  }) = _NftCollectionState;
-}
