--- packages/cryptoplease/lib/features/nft/bl/offchain_metadata_repository.dart
@@ -1,11 +1,0 @@
-import 'package:solana/metaplex.dart';
-
-class OffchainMetadataRepository {
-  final Map<String, Future<OffChainMetadata>> _requests = {};
-
-  Future<OffChainMetadata> getMetadata(Metadata metadata) async =>
-      _requests.putIfAbsent(
-        metadata.mint,
-        () => metadata.getExternalJson(),
-      );
-}
