+++ packages/cryptoplease/lib/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/event.dart
