--- packages/cryptoplease/lib/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/nft/event.dart
@@ -1,19 +1,0 @@
-part of 'bloc.dart';
-
-@freezed
-class NftCreateOutgoingTransferEvent with _$NftCreateOutgoingTransferEvent {
-  const factory NftCreateOutgoingTransferEvent.recipientUpdated(
-    String address,
-  ) = RecipientUpdated;
-
-  const factory NftCreateOutgoingTransferEvent.referenceUpdated(
-    String reference,
-  ) = ReferenceUpdated;
-
-  const factory NftCreateOutgoingTransferEvent.memoUpdated(String memo) =
-      MemoUpdated;
-
-  const factory NftCreateOutgoingTransferEvent.cleared() = Cleared;
-
-  const factory NftCreateOutgoingTransferEvent.submitted() = Submitted;
-}
