+++ packages/cryptoplease/lib/features/outgoing_transfer/presentation/send_flow/confirm_screen/components/amount_fee_view.dart
