+++ packages/cryptoplease/lib/features/outgoing_transfer/presentation/send_flow/confirm_screen/components/amount_view.dart
@@ -2,7 +2,7 @@ import 'package:cryptoplease/core/amount.dart';
 import 'package:cryptoplease/core/conversion_rates/presentation/conversion_rates.dart';
 import 'package:cryptoplease/core/currency.dart';
 import 'package:cryptoplease/core/presentation/format_amount.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/fungible_token/confirm_screen/components/list_item.dart';
+import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/confirm_screen/components/list_item.dart';
 import 'package:cryptoplease/l10n/device_locale.dart';
 import 'package:flutter/material.dart';
