+++ packages/cryptoplease/lib/features/outgoing_transfer/presentation/send_flow/confirm_screen/components/list_item.dart
