+++ packages/cryptoplease/lib/features/outgoing_transfer/presentation/send_flow/confirm_screen/components/token_create_link_content.dart
@@ -1,6 +1,6 @@
 import 'package:cryptoplease/app/components/info_icon.dart';
 import 'package:cryptoplease/core/amount.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/fungible_token/confirm_screen/components/amount_fee_view.dart';
+import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/confirm_screen/components/amount_fee_view.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease_ui/cryptoplease_ui.dart';
 import 'package:flutter/widgets.dart';
