+++ packages/cryptoplease/lib/features/outgoing_transfer/presentation/send_flow/link_transfer_flow_screen.dart
@@ -6,10 +6,10 @@ import 'package:cryptoplease/core/conversion_rates/bl/repository.dart';
 import 'package:cryptoplease/core/split_key_payments/split_key_api_version.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/core/user_preferences.dart';
-import 'package:cryptoplease/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/ft/bloc.dart';
+import 'package:cryptoplease/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/bloc.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/outgoing_payment.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/repository.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/fungible_token/enter_amount_screen.dart';
+import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/enter_amount_screen.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
