+++ packages/cryptoplease/lib/features/outgoing_transfer/presentation/send_flow/routes.dart
@@ -1,13 +1,10 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/common/enter_address_screen.dart';
 import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/common/pick_recipient_type_screen.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/fungible_token/confirm_screen/confirm_screen.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/fungible_token/direct_transfer_flow_screen.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/fungible_token/enter_amount_screen.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/fungible_token/link_transfer_flow_screen.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/non_fungible_token/confirm_screen/confirm_screen.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/non_fungible_token/direct_transfer_flow_screen.dart';
-import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/non_fungible_token/link_transfer_flow_screen.dart';
+import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/confirm_screen/confirm_screen.dart';
+import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/direct_transfer_flow_screen.dart';
+import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/enter_amount_screen.dart';
+import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/link_transfer_flow_screen.dart';
 import 'package:cryptoplease/features/qr_scanner/qr_scanner_request.dart';
 import 'package:cryptoplease/features/qr_scanner/qr_scanner_screen.dart';
 
@@ -28,20 +25,6 @@ const createPaymentFlowRoutes = [
     children: [
       AutoRoute<void>(page: EnterAmountScreen),
       AutoRoute<void>(page: FtConfirmScreen),
-    ],
-  ),
-  AutoRoute<void>(
-    page: NftDirectTransferFlowScreen,
-    children: [
-      AutoRoute<QrScannerRequest>(page: QrScannerScreen),
-      AutoRoute<void>(page: EnterAddressScreen),
-      AutoRoute<void>(page: NftConfirmScreen),
-    ],
-  ),
-  AutoRoute<void>(
-    page: NftLinkTransferFlowScreen,
-    children: [
-      AutoRoute<void>(page: NftConfirmScreen),
     ],
   ),
 ];
