+++ packages/cryptoplease/lib/features/request_pay/bl/request_pay_bloc.dart
@@ -2,7 +2,7 @@ import 'package:cryptoplease/core/amount.dart';
 import 'package:cryptoplease/core/currency.dart';
 import 'package:cryptoplease/core/split_key_payments/split_key_api_version.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
-import 'package:cryptoplease/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/ft/bloc.dart';
+import 'package:cryptoplease/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/bloc.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/outgoing_payment.dart';
 import 'package:decimal/decimal.dart';
 import 'package:dfunc/dfunc.dart';
